% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{num_scale}
\alias{num_scale}
\alias{scale_y_number}
\alias{scale_x_number}
\title{Easy number scales.}
\usage{
scale_y_number(scale = 1, ...)

scale_x_number(scale = 1, ...)
}
\arguments{
\item{scale}{Shorthand number in scientific format, must be either 1, 1e3,1e6, or 1e9 to be recognised. Will be inverted in the scale (1/1e3).}

\item{...}{arguments passed on to ggplot2::scale_y_continuous.}
}
\description{
Quickly format big.marks to commas, and scale numbers to metric units.
}
