% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ziba.R
\name{theme_ziba_ridges}
\alias{theme_ziba_ridges}
\title{Ziba theme for ridge plots}
\usage{
theme_ziba_ridges(font_family = "Roboto")
}
\arguments{
\item{font_family}{Font to use in the graphs, defaults to Roboto.}
}
\value{
A ggplot2 theme
}
\description{
Ridge plots require a different offset for y axis labels, and look better with lighter grid lines.
}
\examples{
# ADD_EXAMPLES_HERE
}
