% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{ziba_gradient}
\alias{ziba_gradient}
\title{Create a blue gradient}
\usage{
ziba_gradient(n)
}
\arguments{
\item{n}{number of shades.}
}
\value{
character vector of \code{n} colours in hexidecimal format.
}
\description{
Extends RColorBrewer with functionality for more than 9 shades of blue.
}
