% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{zibablue}
\alias{zibablue}
\title{A beautiful blue}
\usage{
zibablue()
}
\value{
a colour in hexidecimal format
}
\description{
Returns standard ziba blue for consistency
}
\examples{
# ADD_EXAMPLES_HERE
}
