% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ziba.R
\name{zibasave}
\alias{zibasave}
\title{Save your Zibas plot}
\usage{
zibasave(plt, filename, width = 15, device = "png", family = "Spline Sans")
}
\arguments{
\item{plt}{A ggplot2 plot.}

\item{filename}{Location to save plot.}

\item{width}{Width, in cm.}

\item{device}{Output device, e.g., png, pdf...}

\item{family}{Font to use in the graphs, defaults to Spline Sans.}
}
\description{
Wrapper for ggplot2's save function with presets that work with the font family.
}
\examples{
# ADD_EXAMPLES_HERE
}
