% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ziba.R
\name{theme_ziba}
\alias{theme_ziba}
\title{Ziba Theme}
\usage{
theme_ziba(font_family = "Segoe UI")
}
\arguments{
\item{font_family}{Font to use in the graphs, defaults to Segoe UI}
}
\value{
A ggplot2 theme.
}
\description{
Good looking, based on principles from 'Storytelling with data'.
}
