% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ziba-save.R
\name{zibasave}
\alias{zibasave}
\title{Save your Zibas plot}
\usage{
zibasave(filename, plt = ggplot2::last_plot(), path = ".", pdf = FALSE, ...)
}
\arguments{
\item{filename}{desired name of file on disk}
}
\description{
Wrapper for ggplot2's save function with presets that work with the font family.
}
