% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ziba.R
\name{zibasave}
\alias{zibasave}
\title{Save your Zibas plot}
\usage{
zibasave(
  filename,
  plt = ggplot2::last_plot(),
  path = ".",
  width = 22,
  device = "png",
  family = "Spline Sans",
  ...
)
}
\arguments{
\item{filename}{File name to create on disk.}

\item{plt}{Plot to save, defaults to last plot displayed.}

\item{path}{Path of the directory to save plot to: path and filename are
combined to create the fully qualified file name. Defaults to the working directory.}

\item{width}{Width, in cm.}

\item{device}{Device to use. Can either be a device function (e.g. png), or one
of "eps", "ps", "tex" (pictex), "pdf", "jpeg", "tiff", "png", "bmp", "svg" or "wmf" (windows only).}

\item{family}{Font to use in the graphs, defaults to Spline Sans.}
}
\description{
Wrapper for ggplot2's save function with presets that work with the font family.
}
\examples{
# ADD_EXAMPLES_HERE
}
