% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ziba-save.R
\name{zibasave}
\alias{zibasave}
\title{Save your Zibas plot}
\usage{
zibasave(
  filename = NULL,
  plot = ggplot2::last_plot(),
  path = ".",
  pdf = FALSE,
  bg = "white"
)
}
\arguments{
\item{filename}{desired name of file on disk. If NULL, defaults to plot title, if there is one, otherwise "figure\%03d".}

\item{plot}{The plot to print, defaults to last plot}

\item{pdf}{Use pdf format instead of default png}
}
\description{
Wrapper for ggplot2's save function with presets that work with the font family.
}
